requirements.txt
typer==0.15.1
colorama==0.4.6
shellingham==1.5.4
pytest==8.3.4
python-dotenv==1.0.1